django_platform/requirements.txt
# Django Stock Analysis Platform Requirements

# Core Django
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-ratelimit==4.1.0
django-environ==0.11.2

# Database
psycopg2-binary==2.9.7
redis==5.0.1
django-redis==5.4.0

# Authentication & Security
djangorestframework-simplejwt==5.3.0
cryptography==41.0.7
django-oauth-toolkit==1.7.1

# API & Documentation
drf-yasg==1.21.7
django-filter==23.3
django-extensions==3.2.3

# Background Tasks
celery==5.3.4
django-celery-beat==2.5.0
django-celery-results==2.5.1

# Stock Data APIs
yfinance==0.2.18
alpha-vantage==2.3.1
requests==2.31.0
pandas==2.1.3
numpy==1.25.2

# AI Analysis
groq==0.4.0
openai==1.3.5

# Data Processing
plotly==5.17.0
ta==0.10.2
beautifulsoup4==4.12.2

# WebSocket Support
channels==4.0.0
channels-redis==4.1.0

# Production
gunicorn==21.2.0
whitenoise==6.6.0
sentry-sdk==1.38.0

# Development
django-debug-toolbar==4.2.0
pytest-django==4.6.0
factory-boy==3.3.0
coverage==7.3.2

# Monitoring
django-prometheus==2.3.1
psutil==5.9.6
